chore(utils): renames the ipfs utils file PE-1885
